Create README.md
Kola Pattern Generator makes it easy to create patterns for the Galaxy Hoop. This program creates a list of colors to populate a 16x16 grid. The Length can be considered the LEDs and the height can be considered 'time'. The first row(read left to right #1 - #16) can be considered "Frame1". The second row should be considered "Frame2" and so on.. The LEDs on the Galaxy Hoop will be split up in to groups of 16 during display. Example:

x=black(blank)    r=red    g=green    b=blue

F1-xxrgxxxxxxxxxxxx 
F2-xxrxgxxxxxbxxxbx 
F3-xxrxxgxxxxxxbxxx 
F4-xxrxxxgxxxbxxxbx 
F5-xxrxxxxgxxbbbbbx 
F6-xxrxxxxxgxxxxxxx 

The 'Red' LED in this example will stay in place. The 'Green' LED will move to the right. The 'Blue' LED will create a happy face.

As of right now, some coding is necessary to get the colors to display on the hoop. Goal: Use text injection to create 'ready-to-go' .pde files for Arduino environment.